Moved test_xt1511_policy to common includes to let it get used by the thermocycler tests
